--- sample/src/main/java/com/codenjoy/dojo/sample/model/Elements.java
@@ -1,45 +1,0 @@
-package com.codenjoy.dojo.sample.model;
-
-import com.codenjoy.dojo.services.CharElements;
-
-/**
- * Тут указана легенда всех возможных объектов на поле и их состояний.
- * Важно помнить, что для каждой енумной константы надо создать спрайт в папке \src\main\webapp\resources\sprite.
- */
-public enum Elements implements CharElements {
-
-    // TODO
-    NONE(' '),      // например это пустое место, куда можно перейти герою
-    WALL('☼'),       // а это стенка, через которую я хочу чтобы проходить нельзя было
-    HERO('☺'),       // а это мой герой
-    BULLET('*'),     // то снаряд
-    STONE('0'),     // то астероид
-    EXPLOSION('x'),     // взрыв
-    BOMB('♣');       // бомба
-
-    final char ch;
-
-    Elements(char ch) {
-        this.ch = ch;
-    }
-
-    @Override
-    public char ch() {
-        return ch;
-    }
-
-    @Override
-    public String toString() {
-        return String.valueOf(ch);
-    }
-
-    public static Elements valueOf(char ch) {
-        for (Elements el : Elements.values()) {
-            if (el.ch == ch) {
-                return el;
-            }
-        }
-        throw new IllegalArgumentException("No such element for " + ch);
-    }
-
-}
