--- sample/src/main/java/com/codenjoy/dojo/sample/model/Gold.java
@@ -1,25 +1,0 @@
-package com.codenjoy.dojo.sample.model;
-
-import com.codenjoy.dojo.services.Point;
-import com.codenjoy.dojo.services.PointImpl;
-import com.codenjoy.dojo.services.State;
-
-/**
- * Артефакт Золото на поле
- */
-//public class Gold extends PointImpl implements State<Elements, Player> {
-
-    // TODO
-//    public Gold(int x, int y) {
-//        super(x, y);
-//    }
-//
-//    public Gold(Point point) {
-//        super(point);
-//    }
-//
-//    @Override
-//    public Elements state(Player player, Object... alsoAtPoint) {
-//        return Elements.GOLD;
-//    }
-//}
