--- sample/src/main/java/com/codenjoy/dojo/sample/model/Level.java
@@ -1,20 +1,0 @@
-package com.codenjoy.dojo.sample.model;
-
-import java.util.List;
-
-/**
- * Я вот для простоты и удобства хочу указывать борду в тестовом виде, а реализация этого интерфейса позволяет мне это сделать
- */
-public interface Level {
-
-    /**
-     * @return Размер поля (обязательно квадратное)
-     */
-    int getSize();
-
-    List<Hero> getHero();
-
-    List<Wall> getWalls();
-
-    List<Stone> getStones();
-}
