--- sample/src/main/java/com/codenjoy/dojo/sample/model/Wall.java
@@ -1,20 +1,0 @@
-package com.codenjoy.dojo.sample.model;
-
-import com.codenjoy.dojo.services.Point;
-import com.codenjoy.dojo.services.PointImpl;
-import com.codenjoy.dojo.services.State;
-
-/**
- * Created by indigo on 01.08.2015.
- */
-public class Wall extends PointImpl implements State<Elements, Player> {
-
-    public Wall(Point pt) {
-        super(pt);
-    }
-
-    @Override
-    public Elements state(Player player, Object... alsoAtPoint) {
-        return Elements.WALL;
-    }
-}
