--- sample/src/main/java/com/codenjoy/dojo/sample/services/Events.java
@@ -1,9 +1,0 @@
-package com.codenjoy.dojo.sample.services;
-
-/**
- * Ивенты, которые могут возникать в игре опиши тут. Что есть ивенты? ну убили твоего героя и ты хочешь ему очков начислить штрафных
- * или, быть может, наоборот - он поднял что-то ценное и ты хочешь ему дать бонус. Вот все все ивенты.
- */
-public enum Events {
-//    WIN, LOOSE;
-}
