+++ spacerace/src/main/java/com/codenjoy/dojo/spacerace/client/YourSolver.java
@@ -1,0 +1,48 @@
+package com.codenjoy.dojo.spacerace.client;
+
+import com.codenjoy.dojo.client.Direction;
+import com.codenjoy.dojo.client.Solver;
+import com.codenjoy.dojo.client.WebSocketRunner;
+import com.codenjoy.dojo.services.Dice;
+import com.codenjoy.dojo.services.RandomDice;
+
+/**
+ * User: your name
+ * Это твой алгоритм AI для игры. Реализуй его на свое усмотрение.
+ * Обрати внимание на {@see YourSolverTest} - там приготовлен тестовый
+ * фреймворк для тебя.
+ */
+public class YourSolver implements Solver<Board> {
+
+    private static final String USER_NAME = "user@gmail.com";
+
+    private Dice dice;
+    private Board board;
+
+    public YourSolver(Dice dice) {
+        this.dice = dice;
+    }
+
+    @Override
+    public String get(Board board) {
+        this.board = board;
+//        if (board.isGameOver()) return "";
+
+        return Direction.UP.toString();
+    }
+
+    public static void main(String[] args) {
+        start(USER_NAME, WebSocketRunner.Host.REMOTE);
+    }
+
+    public static void start(String name, WebSocketRunner.Host server) {
+        try {
+            WebSocketRunner.run(server, name,
+                    new YourSolver(new RandomDice()),
+                    new Board());
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+    }
+
+}
