+++ spacerace/src/main/java/com/codenjoy/dojo/spacerace/client/ai/ApofigSolver.java
@@ -1,0 +1,49 @@
+package com.codenjoy.dojo.spacerace.client.ai;
+
+import com.codenjoy.dojo.client.LocalGameRunner;
+import com.codenjoy.dojo.client.Solver;
+import com.codenjoy.dojo.client.WebSocketRunner;
+import com.codenjoy.dojo.spacerace.client.Board;
+import com.codenjoy.dojo.spacerace.services.GameRunner;
+import com.codenjoy.dojo.services.Dice;
+import com.codenjoy.dojo.services.RandomDice;
+
+/**
+ * Это алгоритм твоего бота. Он будет запускаться в игру с первым
+ * зарегистрировавшимся игроком, чтобы ему не было скучно играть самому.
+ * Реализуй его как хочешь, хоть на Random.
+ * Для его запуска воспользуйся методом {@see ApofigSolver#main}
+ */
+public class ApofigSolver implements Solver<Board> {
+
+    public ApofigSolver(Dice dice) {
+        // TODO
+    }
+
+    @Override
+    public String get(final Board board) {
+        // TODO
+        return "UP,ACT(1,0,2),DOWN";
+    }
+
+    /**
+     * Метод для запуска игры с текущим ботом. Служит для отладки.
+     */
+    public static void main(String[] args) {
+        LocalGameRunner.run(new GameRunner(),
+                new ApofigSolver(new RandomDice()),
+                new Board());
+//        start(WebSocketRunner.DEFAULT_USER, WebSocketRunner.Host.LOCAL);
+    }
+
+    public static void start(String name, WebSocketRunner.Host server) {
+        try {
+            WebSocketRunner.run(server, name,
+                    new ApofigSolver(new RandomDice()),
+                    new Board());
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+    }
+
+}
