+++ spacerace/src/main/java/com/codenjoy/dojo/spacerace/model/Gold.java
@@ -1,0 +1,21 @@
+package com.codenjoy.dojo.spacerace.model;
+
+/**
+ * Артефакт Золото на поле
+ */
+//public class Gold extends PointImpl implements State<Elements, Player> {
+
+    // TODO
+//    public Gold(int x, int y) {
+//        super(x, y);
+//    }
+//
+//    public Gold(Point point) {
+//        super(point);
+//    }
+//
+//    @Override
+//    public Elements state(Player player, Object... alsoAtPoint) {
+//        return Elements.GOLD;
+//    }
+//}
