+++ spacerace/src/main/java/com/codenjoy/dojo/spacerace/model/Player.java
@@ -1,0 +1,75 @@
+package com.codenjoy.dojo.spacerace.model;
+
+import com.codenjoy.dojo.spacerace.services.Events;
+import com.codenjoy.dojo.services.EventListener;
+
+/**
+ * Класс игрока. Тут кроме героя может подсчитываться очки. Тут же ивенты передабтся лиснеру фреймворка.
+ */
+public class Player {
+
+    private EventListener listener;
+    private int maxScore;
+    private int score;
+    Hero hero;
+
+    /**
+     * @param listener Это шпийон от фреймоврка. Ты должен все ивенты которые касаются конкретного пользователя сормить ему.
+     */
+    public Player(EventListener listener) {
+        this.listener = listener;
+        clearScore();
+    }
+
+    private void increaseScore() {
+        score = score + 1;
+        maxScore = Math.max(maxScore, score);
+    }
+
+    public int getMaxScore() {
+        return maxScore;
+    }
+
+    public int getScore() {
+        return score;
+    }
+
+    /**
+     * Борда может файрить ивенты юзера с помощью этого метода
+     * @param event тип ивента
+     */
+    public void event(Events event) {
+//        switch (event) {
+//            case LOOSE: gameOver(); break;
+//            case WIN: increaseScore(); break;
+//        }
+
+        if (listener != null) {
+            listener.event(event);
+        }
+    }
+
+    private void gameOver() {
+        score = 0;
+    }
+
+    public void clearScore() {
+        score = 0;
+        maxScore = 0;
+    }
+
+    public Hero getHero() {
+        return hero;
+    }
+
+    /**
+     * Когда создается новая игра для пользователя, кто-то должен создать героя
+     * @param field борда
+     */
+    public void newHero(Field field) {
+//        Point pt = field.getFreeRandom();
+//        hero = new Hero(pt);
+//        hero.init(field);
+    }
+
+}
