+++ spacerace/src/main/java/com/codenjoy/dojo/spacerace/model/Stone.java
@@ -1,0 +1,32 @@
+package com.codenjoy.dojo.spacerace.model;
+
+import com.codenjoy.dojo.services.*;
+
+public class Stone extends PointImpl implements State<Elements, Player>, Tickable {
+    private Direction direction;
+
+    public Stone(int x, int y) {
+        super(x, y);
+        direction = Direction.DOWN;
+    }
+
+    public Stone(Point pt) {
+        super(pt);
+    }
+
+    @Override
+    public Elements state(Player player, Object... alsoAtPoint) {
+        return Elements.STONE;
+    }
+
+    @Override
+    public void tick() {
+        if (direction != null) {
+
+            int newX = direction.changeX(x);
+            int newY = direction.changeY(y);
+            move(newX, newY);
+        }
+
+    }
+}
