+++ spacerace/src/main/java/com/codenjoy/dojo/spacerace/services/Events.java
@@ -1,0 +1,9 @@
+package com.codenjoy.dojo.spacerace.services;
+
+/**
+ * Ивенты, которые могут возникать в игре опиши тут. Что есть ивенты? ну убили твоего героя и ты хочешь ему очков начислить штрафных
+ * или, быть может, наоборот - он поднял что-то ценное и ты хочешь ему дать бонус. Вот все все ивенты.
+ */
+public enum Events {
+//    WIN, LOOSE;
+}
