+++ spacerace/src/main/webapp/resources/sprite/sample/bomb.png
