+++ spacerace/src/main/webapp/resources/sprite/sample/dead_hero.png
