+++ spacerace/src/main/webapp/resources/sprite/sample/gold.png
