+++ spacerace/src/main/webapp/resources/sprite/sample/hero.png
