+++ spacerace/src/main/webapp/resources/sprite/sample/other_hero.png
