+++ spacerace/src/main/webapp/resources/sprite/sample/wall.png
